refactor(cards): refactor cards.py into relevant files
The cards.py file was doing too many things in one spot so I spread it
out a bit.

The Card model is currently a wip so the random card from the cards
route is just a placeholder until I have the data model worked out.

Otherwise, the refactor consists of splitting the database functions in
the database/cards_db.py file and the scryfall api interaction in the
services/scryfall_service.py.

The changes made the the __init__ files are to support the new
architecture.