Cards (#20)
* feat(database.cards): update cards cli and init database on flask start

Add a function with guard to __init__.py to initialize and populate the
cards.db.

Update the cli command cards populate with, -r/--rebuild, -f/--force,
options.

Update the cards_schema to include parsed typeline information.

* refactor(cards): refactor cards.py into relevant files

The cards.py file was doing too many things in one spot so I spread it
out a bit.

The Card model is currently a wip so the random card from the cards
route is just a placeholder until I have the data model worked out.

Otherwise, the refactor consists of splitting the database functions in
the database/cards_db.py file and the scryfall api interaction in the
services/scryfall_service.py.

The changes made the the __init__ files are to support the new
architecture.

* chore: updates

* feat(database): add debug output to cards_db.py and add constraints to cards_schema.sql

* feat(models): CardFace complete

* feat(models.card_components.CardFace): add connection parameters

* feat(models.card_components): add component files

* feat(models.card): add card model

* feat(cards_db): add update_default_cards to populate/rebuild commands

* feat(card_components.wrappers): add Prices

* feat(card_components.wrappers): add to_dict to Prices

* feat(card_components.enums): add Rarity

* feat(card_components.enums): add special/bonus to Rarity

* feat(cards.py): swap to using card_service

* feat(services): add card_service

* feat(models.card_components.card_faces): add error handling

* feat(models.card): add more properties and swap to using card_service

* feat(templates.cards): WIP

* feat(models.card): add more properties, swap to CardService for db
interaction

* fix(models.card): add missing argument for defense property

* feat(models.card_components.card_faces): add CardService, modify
properties

* feat(services.card_service): add CardService

* fix(services.card_service): add quotes to Card return value type hints

* feat(templates.cards): WIP